Remove unused ui files
Remove the placeholder ui components
