+++ src/presentation/app/app.component.html
@@ -2,4 +2,4 @@
   [initializeEngine]="siteDataReady"
   [siteContent]="site"
 ></app-platform>
-<app-ui></app-ui>
+<sourcecode-link></sourcecode-link>
