+++ src/presentation/app/ui/sourcecode-link/sourcecode-link.component.html
@@ -1,0 +1,3 @@
+<!-- <div *ngIf="showGithubBanner" class="github-banner-corner">
+  <a [href]="githubBannerUrl" target="_blank" rel="noopener noreferrer">View on GitHub</a>
+</div> -->
