--- src/presentation/app/ui/ui-infobar-bottom/ui-infobar-bottom.component.html
@@ -1,3 +1,0 @@
-<p>
-  ui-infobar-bottom works!
-</p>
