--- src/presentation/app/ui/ui-infobar-bottom/ui-infobar-bottom.component.ts
@@ -1,20 +1,0 @@
-import { ChangeDetectionStrategy, Component, OnInit, ViewEncapsulation } from '@angular/core';
-
-@Component({
-  selector: 'app-ui-infobar-bottom',
-  standalone: true,
-  encapsulation: ViewEncapsulation.None,
-  changeDetection: ChangeDetectionStrategy.OnPush,
-  templateUrl: './ui-infobar-bottom.component.html'
-})
-export class UiInfobarBottomComponent implements OnInit {
-
-  public constructor() { 
-    console.debug('UiInfobarBottomComponent constructor called');
-  }
-
-  public ngOnInit(): void {
-    console.debug('UiInfobarBottomComponent ngOnInit called');
-  }
-
-}
