--- src/presentation/app/ui/ui-infobar-top/ui-infobar-top.component.html
@@ -1,3 +1,0 @@
-<p>
-  ui-infobar-top works!
-</p>
