--- src/presentation/app/ui/ui-sidebar-left/ui-sidebar-left.component.html
@@ -1,3 +1,0 @@
-<p>
-  ui-sidebar-left works!
-</p>
