--- src/presentation/app/ui/ui-sidebar-right/ui-sidebar-right.component.html
@@ -1,3 +1,0 @@
-<p>
-  ui-sidebar-right works!
-</p>
