--- src/presentation/app/ui/ui.component.html
@@ -1,6 +1,0 @@
-<div class="ui-wrapper" *ngIf="false">
-  <app-ui-infobar-top></app-ui-infobar-top>
-  <app-ui-sidebar-left></app-ui-sidebar-left>
-  <app-ui-sidebar-right></app-ui-sidebar-right>
-  <app-ui-infobar-bottom></app-ui-infobar-bottom>
-</div>
